"FastHooks"
this is my playground implementation of a different plugin hook system. I implemented it for few of the most common plugin hooks used in HOT.

the problem with the current plugin system in HOT is that it makes a lot of work in `run`. This work should be done in `add`, which acts as a kind of "precompiler" of hooks.

In "FastHooks", the equivalent of `add` function is called `after`. It works by function chaining, which I would say is the simplest "precompilation" I can think of.

I *was* very excited about this change, but actually on my hardware, I don't notice any significant difference. My gut feeling says that there must be a difference, though.

I apologize showing my point with a new, a little bit too naive class. If you want to use this idea, my recommendation is to reimplement it in the actual PluginHooks class, with full backwards compatibility. It is possible to implement function chaining with support for the current concepts: local hooks, global hooks, once hooks - just keep in mind to do all the work in `add` and no work in `run`.